+++ api/core/src/modules/activity/activity.module.ts
@@ -1,10 +1,11 @@
 import { Module } from "@nestjs/common";
 import { ActivityController } from "./activity.controller";
+import { ActivityRepository } from "./activity.repository";
 import { ActivityService } from "./activity.service";
 
 @Module({
   controllers: [ActivityController],
-  providers: [ActivityService],
+  providers: [ActivityRepository, ActivityService],
   exports: [ActivityService]
 })
 export class ActivityModule {}